requirements.txt
pandas
scikit-learn
tqdm
kagglehub
matplotlib
numpy
torch
torchvision
Pillow